nrf_security: CRACEN: Decrease flash memory overhead of Cracen
This commit is aimed to decrease Flash memory overhead of Cracen,
especially after cracen_psa_key_derivation refactoring.

It also contains minor fixes for the Cracen driver.

Reason: flash overflow in sample.mcuboot_with_encryption for
nRF54LM20A.

Ref: NCSDK-37768

+++ subsys/nrf_security/src/drivers/cracen/cracenpsa/src/internal/key_derivation/cracen_kdf_common.c
@@ -62,8 +62,7 @@ psa_status_t cracen_kdf_generate_output_bytes(cracen_key_derivation_operation_t 
 			/* Copy out buffered output. This may be a partial
 			 * block.
 			 */
-			unsigned int out =
-				MIN(output_length, operation->output_block_available_bytes);
+			size_t out = MIN(output_length, operation->output_block_available_bytes);
 
 			memcpy(output, operation->output_block, out);
